+++ Projects/Weather_Station/Src/es_wifi.c
@@ -1636,7 +1636,7 @@ ES_WIFI_Status_t ES_WIFI_SendData(ES_WIFIObject_t *Obj, uint8_t Socket, uint8_t 
   * @param  len : pointer to the length of the data to be received
   * @retval Operation Status.
   */
-ES_WIFI_Status_t ES_WIFI_ReceiveData(ES_WIFIObject_t *Obj, uint8_t Socket, float *pdata, uint16_t Reqlen, uint16_t *Receivedlen, uint32_t Timeout)
+ES_WIFI_Status_t ES_WIFI_ReceiveData(ES_WIFIObject_t *Obj, uint8_t Socket, uint8_t *pdata, uint16_t Reqlen, uint16_t *Receivedlen, uint32_t Timeout)
 {
   ES_WIFI_Status_t ret = ES_WIFI_STATUS_ERROR;
